moved viewmodel to correct package
